Fix segfault on program launch
The GtkBuilder used to construct the main window gets passed to the View, but the View did not increment the reference count of the object, so when create_scroom() unref-ed the object when it finished, the program would crash. The View class now increments the reference count in its constructor and decrements it in its destructor.